+++ ui/src/components/tip/QasTip.vue
@@ -5,8 +5,6 @@
 </template>
 
 <script setup>
-import { Spacing, SpacingWithUnit } from '../../enums/Spacing'
-
 import { computed } from 'vue'
 
 defineOptions({ name: 'QasTip' })
@@ -24,8 +22,7 @@ const props = defineProps({
 
   size: {
     type: String,
-    default: Spacing.Md,
-    validator: value => Object.values(Spacing).includes(value)
+    default: '16px'
   },
 
   color: {
@@ -36,29 +33,14 @@ const props = defineProps({
 
 // computeds
 const iconProps = computed(() => {
-  const spacingKey = getNormalizedSpacingKey(props.size)
-
   return {
     name: props.icon,
-    size: SpacingWithUnit[spacingKey],
+    size: props.size,
     color: props.color
   }
 })
-
-// functions
-/**
- * Necessário para encontrar a chave correta no enum SpacingWithUnit.
- *
- * @param spacing {string} - spacing ('xs', 'sm', 'md', 'lg', 'xl', '2xl', '3xl', '4xl', '5xl')
- */
-function getNormalizedSpacingKey (spacing) {
-  // Normaliza a primeira letra (ex: 'md' => 'Md')
-  const capitalized = spacing.charAt(0).toUpperCase() + spacing.slice(1)
-
-  // Valida se existe a chave "Md" no enum
-  if (Object.prototype.hasOwnProperty.call(SpacingWithUnit, capitalized)) return capitalized
-}
 </script>
+
 <style lang="scss">
 .qas-tip {
   cursor: help;